qcacld-3.0: Remove logic to disconnect STA on VDEV delete
With Vdev delete not using serialization, no need to disconnect
any STA with connection in progress.

The logic was added to remove the active connect command from
serialization queue so that vdev delete can complete before hdd
time out.

Now as vdev delete is not using the serialization remove this.

Change-Id: Id8c5ab31e54c8cf26999aff58c0773213f0cb13a
CRs-Fixed: 2585753
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>