qcacld-3.0: Enable transition of STA from 20->40, 80 Mhz
Scenario: Turn on SAP in 2.4ghz in 20Mhz and connect
DUT to it.
Now trigger channel change command and change the SAP
channel to 5ghz in 80Mhz mode.

Issue:-
STA would not go to 80Mhz mode and would remain in 20Mhz
Issue in driver was that it was checking for the ht
supported chan width rather than the self capability.
HT self capability would always be 0 if the connection
is in 20Mhz, and thus the STA would never be able to
ramp up it's BW.

Expectation:
STA should move to 80mhz if it supports 80Mhz bandwidth.

Fix:-
Check for the self capability rather than the ht channel
BW supported.

Change-Id: I546d5edd9881c1013b1a08989cbab3d4ef8fae4d
CRs-Fixed: 2594982
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>