qcacld-3.0: Remove the check for 0 scan results
Currently the logic in ACS for 0 scan results is
to select default channel from the ACS channel list,
but the available channels may not always be capable
of the max BW sent by ACS, for example in HW MODE as
ANY then all the channels 2.4ghz and 5ghz would be
present in the ACS channel list, 2.4ghz channels at
the first followed by the 5ghz channels, and if the
scan results are 0 then 2.4ghz channel would be
selected as default which should not be the case.

Fix is to remove the check to select default channel
and let me ACS algorithm decide the best channel
accounting other factors such as noise floor
and max tx rx clear count, power also.

Change-Id: I21e5d73e12fffb00c4996c406d62361b74db629f
CRs-Fixed: 2601141
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>