qcacld-3.0: Set peer params after vdev restart for SAP
Scenario:
1. Turn on SAP on channel 1 on 2.4ghz 20Mhz BW
2. Connect a STA to it.
3. Change the SAp channel to 36 80mhz BW

Observation:
STA changes the channel BW to 80 but SAP does not
send th packets in 80Mhz to the SAP.

Expectation: SAP should send the packets to STA on 80Mhz.

Issue:
SAP does not set the peer params to the FW for the peer
hence the FW keeps on sending the packets with 20Mhz
which was previously configured to it.

Fix:
Send peer params to FW for each peer so that the
packets can be sent by SAP on a max supported BW.

Change-Id: Ic9b310e67001d55b2b5427db0c2ab897ea36b6b0
CRs-Fixed: 2606872
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>