qca-wifi: Add 6GHz support for chan/freq/band usage in mesh mode
Currently mesh mode uses channel numbering scheme which is derived
from APIs which don't support 6GHz channel numbering. This is
important because the channel numbers for 6GHz are overloaded with
those in the 2.4GHz and 5GHz band.

Add support to obtain the correct channel number (and auxiliary
information like band and frequency) through the new APIs that
support 6GHz.

Change-Id: I60afa5294af5e5681974f1bbc45c9e008eda1232
CRs-Fixed: 2605229
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>