qcacld-3.0: Enable MCL feature flag MAX_ALLOC_PAGE_SIZE
This feature flag is used to set the page size to 4KB
for MCL and to use the max_alloc_size config ini param
for page size in the WIN specific code.

The page size is used to allocate link descriptor pages
for the WBM idle link ring instead of the desc bank
contiguous memory allocations.

Change-Id: I9ff575efc475473654cb4b66a95285c779599b24
CRs-Fixed: 2612385
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>