qcacld-3.0: Update peer mac addr before set WEP key
In WEP SAP mode, the ping between SAP and client is failed
because in the SAP side the firmware has no unicast key
for peer mac address. The default address in key context is
SAP bssid.
Update the peer mac address before set key to firmware for
the WEP peer.

Change-Id: Ia5ba4612d0cc0f9900c72baf1c37bd278b526389
CRs-Fixed: 2599289
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>