qcacld-3.0: Conditionally enable WEXT support
Currently CONFIG_WLAN_WEXT_SUPPORT_ENABLE is unconditionally enabled
in many defconfigs, including the default defconfig, but this leads to
build failures when WEXT is not configured in the Kernel. So update
the defconfigs to enable CONFIG_WLAN_WEXT_SUPPORT_ENABLE only when
CONFIG_WIRELESS_EXT is enabled in the Kernel.

Change-Id: Iecfd4e899c2b09c489547dd1a8ff70f71983853f
CRs-Fixed: 2620792
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>