qcacld-3.0: stop the idle restart timer before reading the driver status
If the idle restart is not stopped before reading the driver status
the driver status could be changed in the other thread leading to
different issues.

Stop the idle restart timer before reading the driver status, so
the next action depending on the driver status can be taken correctly.

Change-Id: Ida7fe3fe8144e300344d29cae8188b64f074b3bb
CRs-Fixed: 2626500
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>