qcacmn: Add dummy NLA parsing in driver for kernel >= 5.2
According to new changes in kernel 5.2 version onwards, kernel takes
care of validating the NLA policies as provided by the driver. In this
case there is no need to re-validate the same in the driver.

Add a stub to return true for NLA parse in case the kernel version is
greater than equal to 5.2

Change-Id: Iea1fdcac3000ede2f8b9398cb85443e545cdd3f0
CRs-Fixed: 2632534
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>