qcacld-3.0: Properly handle interface down in case of SSR
Presently in case there is an interface down during SSR, the request is
rejected outright. This causes the driver and the userspace to go out of
sync on the status of that particular interface.

The root cause of this issue is that while SSR is ongoing if interface
down comes, the DSC control via __dsc_vdev_can_trans rejects the interface
down citing QDF_STATUS_E_INVAL. This prevents the request to be queued and
processed later.

To fix this remove the check for driver recovering from the DSC control.

Change-Id: I9598c4606984f924d63e8c459ded0520d0824d08
CRs-Fixed: 2658597
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>